third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/properties/scroll-snap-align-expected.txt

This is a testharness.js-based test.
PASS Can set 'scroll-snap-align' to CSS-wide keywords
PASS Can set 'scroll-snap-align' to var() references
FAIL Can set 'scroll-snap-align' to the 'none' keyword assert_class_string: expected "[object CSSStyleValue]" but got "[object CSSKeywordValue]"
FAIL Can set 'scroll-snap-align' to the 'start' keyword assert_class_string: expected "[object CSSStyleValue]" but got "[object CSSKeywordValue]"
FAIL Can set 'scroll-snap-align' to the 'end' keyword assert_class_string: expected "[object CSSStyleValue]" but got "[object CSSKeywordValue]"
FAIL Can set 'scroll-snap-align' to the 'center' keyword assert_class_string: expected "[object CSSStyleValue]" but got "[object CSSKeywordValue]"
PASS Setting 'scroll-snap-align' to a length throws TypeError
PASS Setting 'scroll-snap-align' to a percent throws TypeError
PASS Setting 'scroll-snap-align' to a time throws TypeError
PASS Setting 'scroll-snap-align' to an angle throws TypeError
PASS Setting 'scroll-snap-align' to a flexible length throws TypeError
PASS Setting 'scroll-snap-align' to a number throws TypeError
PASS Setting 'scroll-snap-align' to a position throws TypeError
PASS Setting 'scroll-snap-align' to a URL throws TypeError
PASS Setting 'scroll-snap-align' to a transform throws TypeError
PASS 'scroll-snap-align' does not supported 'none center'
PASS 'scroll-snap-align' does not supported 'end start'
Harness: the test ran to completion.
